+++ robot_descriptions/tiago_description.py
@@ -6,10 +6,20 @@
 
 """TIAGo description."""
 
+import warnings as _warnings
 from os import getenv as _getenv
 from os import path as _path
 
 from ._cache import clone_to_cache as _clone_to_cache
+
+_warnings.warn(
+    "robot_descriptions.tiago_description is deprecated and will "
+    "switch to the official TIAGo model in robot_descriptions.py in a "
+    "later release. Use robot_descriptions.tiago_official_description "
+    "now to migrate early.",
+    FutureWarning,
+    stacklevel=2,
+)
 
 REPOSITORY_PATH: str = _clone_to_cache(
     "example-robot-data",